Fixes some assembly runtimes

+++ code/modules/assembly/infrared.dm
@@ -201,7 +201,7 @@
 /obj/effect/beam/i_beam/process()
 	//world << "i_beam \ref[src] : process"
 
-	if((loc.density || !(master)))
+	if(loc && loc.density || !master)
 	//	world << "beam hit loc [loc] or no master [master], deleting"
 		qdel(src)
 		return
